split fire-pdf async client
