+++ apps/api/src/controllers/v2/types.ts
@@ -610,7 +610,7 @@ const baseScrapeOptions = z.strictObject({
   __experimental_omceDomain: z.string().optional(),
   __experimental_engpicker: z.boolean().prefault(false).optional(),
   __forceFirePDF: z.boolean().prefault(false).optional(),
-  enableFirePdfAsync: z.boolean().prefault(false).optional(),
+  __experimental_firePdfAsync: z.boolean().prefault(false).optional(),
 });
 
 type ScrapeOptionsBase = z.infer<typeof baseScrapeOptions>;
